Add tests (#12)
* repository done

* util pckg done

* ui test done

* Added services tests

* fixed packages

* Add tests for types

* Added types tests

* preload test

* factory coverage

* added unit tests

* update makefile

* gitignore

+++ internal/service/delete_service.go
@@ -22,7 +22,10 @@ func NewDeleteService(repo *repository.InMemoryRepository, cmd *cobra.Command) *
 }
 
 func (s *DeleteService) DeleteResource(kind, name, namespace string) (string, error) {
-	nKind, _ := util.NormalizeKind(kind)
+	nKind, err := util.NormalizeKind(kind)
+	if err != nil {
+		return "", err
+	}
 	if nKind == "" {
 		return "", fmt.Errorf("you must specify only one resource")
 	}
@@ -31,17 +34,17 @@ func (s *DeleteService) DeleteResource(kind, name, namespace string) (string, er
 		return "", fmt.Errorf("resource(s) were provided, but no name was specified")
 	}
 
-	if namespace == "" {
+	if namespace == "" && kind != "namespace" {
 		return "", fmt.Errorf("a resource cannot be retrieved by name across all namespaces")
-	}
-
-	if nKind == "namespace" {
-		return s.repo.Delete("all", "", name)
 	}
 
 	resources, err := s.repo.List("all", "", "")
 	if err != nil {
 		return "", err
+	}
+
+	if nKind == "namespace" && len(resources) > 0 {
+		return s.repo.Delete("all", "", name)
 	}
 
 	var deletedResources []string
